+++ src/wincalc/model/AreaTriangl.java
@@ -6,7 +6,7 @@ import wincalc.Wincalc;
 
 public class AreaTriangl extends AreaSimple {
 
-    public AreaTriangl(Wincalc iwin, AreaSimple owner, int id, TypeElem typeElem, LayoutArea layout, float width, float height, int color1, int color2, int color3, String param) {
+    public AreaTriangl(Wincalc iwin, AreaSimple owner, float id, TypeElem typeElem, LayoutArea layout, float width, float height, int color1, int color2, int color3, String param) {
         super(iwin, owner, id, typeElem, layout, width, height, color1, color2, color3);
         parsing(param);
     }
